requirements.txt
pytest
requests
openai>=1.0.0
pyyaml
